refine no-install-development.
codegen_src symlink causes following error.

```
Warning: /home/akr/coq/codegen/src was previously bound to codegen_src; it is
remapped to codegen [overriding-logical-loadpath,loadpath]
```

+++ Makefile
@@ -1,4 +1,4 @@
-.PHONY: all install plugin check checker clean no-install-development
+.PHONY: all install plugin check checker clean no-install-development revert-no-install-development
 
 -include Makefile.coq.conf
 
@@ -9,11 +9,28 @@ install : Makefile.coq
 	test ! theories -ef $(COQMF_COQLIB)/user-contrib/codegen || ( echo "$(COQMF_COQLIB)/user-contrib/codegen is linked to theories.  No need to install."; false )
 	$(MAKE) -f Makefile.coq $@
 
-no-install-development :
-	rm -f $(COQMF_COQLIB)/user-contrib/codegen
-	rm -f $(COQMF_COQLIB)/user-contrib/codegen_src
+no-install-development : revert-no-install-development
+ifdef COQMF_COQLIB
+	ln -s ../src/codegen_plugin.cmi theories
+	ln -s ../src/codegen_plugin.cmxs theories
+	ln -s ../src/codegen_plugin.cmxa theories
+	ln -s ../src/codegen_plugin.cmx theories
 	ln -s `pwd`/theories $(COQMF_COQLIB)/user-contrib/codegen
-	ln -s `pwd`/src $(COQMF_COQLIB)/user-contrib/codegen_src
+else
+	-echo COQMF_COQLIB not defined.  run "make" first.
+endif
+
+revert-no-install-development :
+ifdef COQMF_COQLIB
+	rm -f theories/codegen_plugin.cmi
+	rm -f theories/codegen_plugin.cmxs
+	rm -f theories/codegen_plugin.cmxa
+	rm -f theories/codegen_plugin.cmx
+	rm -f $(COQMF_COQLIB)/user-contrib/codegen
+else
+	-echo COQMF_COQLIB not defined.  run "make" first.
+endif
+
 
 plugin : Makefile.coq
 	$(MAKE) -f Makefile.coq src/codegen_plugin.cmxs
